Move PeerJS stuff into settings context
